refactor: keep domain profile loading bounded and below smell limits
